fix: fix up error handling in wireguard
Previously what I was seeing happen is something would error out in the killswitch while adding the exclusions which would cause the script to immediately exit and since the interface was left up the next time it would start up we would get the `wg-quick: `wg0' already exists` message. I've moved the trap up earlier in the execution and have it listening on `EXIT`, so any sort of exit will go through the shutdown. So if the interface was brought up and we die in the killswitch it will still shutdown the interface.

Then as far as my issue in the killswitch it was while adding the exclusions I seem to be hitting some sort of conflict with one of the CIDR's i was adding. But I think instead of dying, I think we should just log a warning that the exclusion couldn't be added and then continue on.

FWIW here's what I was hitting:
```
[INFO] Excluding 10.244.0.0/16 from VPN IPv4 traffic
RTNETLINK answers: File exists
```

This will now log the following and continue on instead of dying:

```
[WARNING] Received non-zero exit code adding route 10.244.0.0/16 via 10.244.0.1
```

Signed-off-by: Branden Cash <ammmze@gmail.com>